add a Makefile. bump versions to 0.2.0. refactor version modules
