+++ openedx_plugin/__about__.py
@@ -7,4 +7,4 @@ date:           feb-2022
 
 usage:          semantic version control for openedx_plugin
 """
-__version__ = "0.1.3"
+__version__ = "0.2.0"
