+++ openedx_plugin/apps.py
@@ -139,7 +139,7 @@ class CustomPluginConfig(AppConfig):
             return
 
         from . import signals  # pylint: disable=unused-import
-        from .version import __version__
+        from .__about__ import __version__
         from .waffle import waffle_init
         from .utils import PluginJSONEncoder
