+++ openedx_plugin_api/apps.py
@@ -62,7 +62,7 @@ class CustomPluginAPIConfig(AppConfig):
             return
 
         from . import signals  # pylint: disable=unused-import
-        from .version import __version__
+        from .__about__ import __version__
         from .waffle import waffle_init
 
         log.info("{label} {version} is ready.".format(label=self.label, version=__version__))
