+++ openedx_plugin_cms/__about__.py
@@ -1,0 +1,2 @@
+# coding=utf-8
+__version__ = "0.2.0"
