--- openedx_plugin_cms/version.py
@@ -1,2 +1,0 @@
-# coding=utf-8
-__version__ = "0.1.1"
