+++ pyproject.toml
@@ -24,7 +24,7 @@ build-backend = "setuptools.build_meta:__legacy__"
 #------------------------------------------------------------------------------
 [project]
 name = "openedx-plugin-example"
-version = "0.1.3"
+version = "0.2.0"
 authors = [
   { name="Lawrence McDaniel", email="lpm0073@gmail.com" }
 ]
